refactor(66939998): create folder ViewModels to separate models
