+++ Project/Project/Controllers/AuthController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Project.Models;
+using Project.Models.ViewModels;
 using Newtonsoft.Json;
 using static BCrypt.Net.BCrypt;
 using Microsoft.EntityFrameworkCore;
