+++ Project/Project/Controllers/HomeController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using Project.Models;
+using Project.Models.ViewModels;
 using Project.Data;
 
 namespace Project.Controllers;
