+++ Project/Project/Controllers/ProfileController.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
 using Project.Data;
 using Project.Models;
+using Project.Models.ViewModels;
 
 namespace Project.Controllers
 {
