+++ Project/Project/Controllers/UserController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Project.Models;
+using Project.Models.ViewModels;
 using Project.Data;
 using Newtonsoft.Json;
 using Microsoft.EntityFrameworkCore;
