+++ Project/Project/Models/ViewModels/ErrorViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Project.Models;
+﻿namespace Project.Models.ViewModels;
 
 public class ErrorViewModel
 {
