+++ Project/Project/Models/ViewModels/LoginViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace Project.Models
+namespace Project.Models.ViewModels
 {
     public class LoginViewModel
     {
