+++ Project/Project/Models/ViewModels/ProfileViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Project.Models
+﻿namespace Project.Models.ViewModels
 {
     public class ProfileViewModel
     {
