+++ Project/Project/Project.csproj
@@ -27,6 +27,7 @@
     <None Remove="HostedServices\" />
     <None Remove="Views\Profile\" />
     <None Remove="Views\User\" />
+    <None Remove="Models\ViewModels\" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Migrations\" />
@@ -37,6 +38,7 @@
     <Folder Include="HostedServices\" />
     <Folder Include="Views\Profile\" />
     <Folder Include="Views\User\" />
+    <Folder Include="Models\ViewModels\" />
   </ItemGroup>
   <ItemGroup>
     <Content Remove="wwwroot\lib\fontawesome\" />
