+++ Project/Project/Views/Auth/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model LoginViewModel
+﻿@model Project.Models.ViewModels.LoginViewModel
 
 @{
     ViewData["Title"] = "Login";
