+++ Project/Project/Views/Profile/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model Project.Models.ProfileViewModel
+﻿@model Project.Models.ViewModels.ProfileViewModel
 
 @{
     ViewData["Title"] = "My Space";
