+++ Project/Project/Views/Shared/Error.cshtml
@@ -1,4 +1,4 @@
-﻿@model ErrorViewModel
+﻿@model Project.Models.ViewModels.ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
