+++ Project/Project/Views/User/Update.cshtml
@@ -1,4 +1,4 @@
-﻿@model AccountViewModel
+﻿@model Project.Models.ViewModels.AccountViewModel
 
 @{
     ViewData["Title"] = "Account";
